requirements.txt
requests
beautifulsoup4
pillow
tqdm
pytesseract
